backendV2/requirements.txt
fastapi[standard]
python-dotenv
torch
torchvision
albumentations
timm
opencv-python
numpy
scipy
scikit-image
pytesseract
Pillow
ultralytics
selenium
webdriver-manager
google-generativeai